Added next pageExtensions config, renamed existing pages
