+++ next.config.js
@@ -4,6 +4,7 @@ const { i18n } = require('./next-i18next.config');
 const nextConfig = {
   i18n,
   reactStrictMode: true,
+  pageExtensions: ['page.tsx', 'page.ts'],
 
   webpack(config, { isServer, dev }) {
     config.experiments = {
