+++ pages/admin/index.page.ts
@@ -1,0 +1,1 @@
+export { default, getServerSideProps } from './users/index.page';
