--- pages/admin/index.ts
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from './users';
