+++ pages/api/auth/[...nextauth].page.ts
