+++ pages/api/google.page.ts
