+++ pages/api/home/home.tsx
@@ -20,7 +20,7 @@ import HomeContext from './home.context';
 import { HomeInitialState, initialState } from './home.state';
 
 import { v4 as uuidv4 } from 'uuid';
-import { authOptions } from '../auth/[...nextauth]';
+import { authOptions } from '../auth/[...nextauth].page';
 import { getServerSession } from 'next-auth/next';
 
 interface Props {
