+++ pages/api/plugins.page.ts
